requirements.txt
django
mysqlclient
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow